@@ -1,10 +1,6 @@
 <?php
 
-/**
- * @see       https://github.com/laminas/laminas-tag for the canonical source repository
- * @copyright https://github.com/laminas/laminas-tag/blob/master/COPYRIGHT.md
- * @license   https://github.com/laminas/laminas-tag/blob/master/LICENSE.md New BSD License
- */
+declare(strict_types=1);
 
 namespace LaminasTest\ServiceManager;
 
@@ -13,6 +9,7 @@ use Laminas\ServiceManager\Test\CommonPluginManagerTrait;
 use Laminas\Tag\Cloud\Decorator\DecoratorInterface;
 use Laminas\Tag\Cloud\DecoratorPluginManager;
 use PHPUnit\Framework\TestCase;
+use RuntimeException;
 
 /**
  * Example test of using CommonPluginManagerTrait
@@ -21,16 +18,19 @@ class DecoratorPluginManagerCompatibilityTest extends TestCase
 {
     use CommonPluginManagerTrait;
 
+    /** @return DecoratorPluginManager */
     protected function getPluginManager()
     {
         return new DecoratorPluginManager(new ServiceManager());
     }
 
+    /** @return string */
     protected function getV2InvalidPluginException()
     {
-        return \RuntimeException::class;
+        return RuntimeException::class;
     }
 
+    /** @return string */
     protected function getInstanceOf()
     {
         return DecoratorInterface::class;
